Move search to app layer
