+++ taxel-gui/src/lib.rs
@@ -5,7 +5,7 @@ pub mod ui;
 
 use anyhow::{Context, Result};
 pub use app::TaxelApp;
-pub use domain::fact_table::{populate_fact_table, FactRow, FactSection, FactTable, SearchHit};
+pub use domain::fact_table::{populate_fact_table, FactRow, FactSection, FactTable};
 pub use infrastructure::report_store;
 use log::debug;
 use std::path::{Path, PathBuf};
